Fixed comments
