+++ src/Uno.UI/UI/Xaml/FontFamily.cs
@@ -11,9 +11,7 @@ namespace Windows.UI.Xaml.Media
 		{
 			Source = familyName;
 
-#if __WASM__
-			ParsedSource = ParseFontFamilySource(familyName);
-#endif
+			Init(familyName);
 
 			// This instance is immutable, we can cache the hash code.
 			_hashCode = Source.GetHashCode();
@@ -58,50 +56,6 @@ namespace Windows.UI.Xaml.Media
 			return ReferenceEquals(a, null) || !a.Equals(b);
 		}
 
-#if __WASM__
-		/// <summary>
-		/// Contains the parsed font family for use in WASM
-		/// (matches CSS @font-face's font-family)
-		/// </summary>
-		internal string ParsedSource { get; }
-
-		private string ParseFontFamilySource(string familyName)
-		{
-			const string ForwardSlash = "/";
-			const string Hash = "#";
-			const string Dot = ".";
-			if (string.IsNullOrEmpty(familyName))
-			{
-				throw new ArgumentException("Font family name must not be empty string nor null", nameof(familyName));
-			}
-			//check if family name is a pure name or a path
-			if (familyName.Contains(ForwardSlash) || familyName.Contains(Hash))
-			{
-				//we have a path to font family name, parse just the name itself
-				//there are two possible formats:
-				//1) "some/path/to/font/MyNiceFont.ttf#My Nice Font" (actually works even with pure "MyNiceFont.ttf#My Font")
-				//   -> we extract the part after #
-
-				var hashFontNameStart = familyName.LastIndexOf(Hash);
-				if (hashFontNameStart != -1)
-				{
-					return familyName.Substring(hashFontNameStart + 1);
-				}
-
-				//or 
-				//2) "some/path/to/font/MyNiceFont.ttf"
-				//   -> we fall back to the font file name
-
-				var slashFontNameStart = familyName.LastIndexOf(ForwardSlash) + 1; //works even if slash is not present at all -> 0				
-				var extensionStart = familyName.LastIndexOf(Dot);
-				if (extensionStart < slashFontNameStart) //no dot after slash
-				{
-					extensionStart = familyName.Length;
-				}
-				return familyName.Substring(slashFontNameStart, extensionStart - slashFontNameStart);
-			}
-			return familyName;
-		}
-#endif
+		partial void Init(string fontName);
 	}
 }
